requirements.txt
# Core numerical stack
numpy>=1.24.0
scipy>=1.11.0

# Data validation and serialization
pydantic>=2.0.0
pyyaml>=6.0

# Geometry operations (2D)
shapely>=2.0.0

# Meshing
gmsh>=4.11.0
pygmsh>=7.1.0
meshio>=5.3.0

# Visualization
matplotlib>=3.7.0
pyvista>=0.42.0

# Development tools
pytest>=7.4.0
pytest-cov>=4.1.0
hypothesis>=6.82.0

# Code quality
ruff>=0.1.0
mypy>=1.5.0

# Utilities
rich>=13.5.0
joblib>=1.3.0

# HDF5 storage (for results)
h5py>=3.9.0
